Refactored tests for side panel
